+++ ouroboros-network-framework/CHANGELOG.md
@@ -9,6 +9,8 @@
 * Updated code to accommodate changes on `PeerSharing` data type.
 * Fixed Server2 [sim test](https://github.com/input-output-hk/ouroboros-network/issues/4607) by synchronizing connection/disconnection events.
 * Changed connection manager `readState` function to be in `STM` instead of `m`
+* Added `waitForOutboundDemotion` function to Connection
+  Manager's API
 
 ## 0.10.0.01 -- 2023-11-02